feat: Update 60-custom.just file
This commit adds a new script `install-pdm` to the `60-custom.just` file. The `install-pdm` script installs pdm, a modern Python package and dependency manager, by downloading and executing a Python script. The script also checks the integrity of the downloaded file using its SHA256 hash. The installation path for pdm is set to the user's home directory.